Changed getCategoryFuzzy(...) to read from the CategorizedTransaction database table instead of the DescriptionCategory database table. This helps to match transactions with descriptions that were previously categorized to that same category, rather than prompting the user to select the category over and over again. Ultimately, I think the DescriptionCategory table can be entirely removed, but that's for a future refactor

+++ src/main/java/ca/jonathanfritz/ofxcat/OfxCat.java
@@ -67,7 +67,9 @@ public class OfxCat {
             if (!Files.isDirectory(backupPath.getParent())) {
                 Files.createDirectories(backupPath.getParent());
             }
-            Files.copy(pathToImportFile, backupPath);
+            if (!Files.exists(backupPath)) {
+                Files.copy(pathToImportFile, backupPath);
+            }
         } catch (IOException ex) {
             throw new CliException("Failed to copy imported file to data directory", ex);
         }
